pyproject.toml
[project]
name = "fit-ctf-virt"
version = "0.1.0"
description = ""
authors = [
    {name = "Hung Do",email = "hungdojan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "invoke (>=2.2.1,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
